+++ sdk/servicebus/azure-servicebus/azure/servicebus/_common/message.py
@@ -11,6 +11,8 @@ import logging
 import copy
 from typing import Optional, List, Union, Iterable, TYPE_CHECKING, Any
 
+import six
+
 import uamqp.errors
 import uamqp.message
 from uamqp.constants import MessageState
@@ -45,7 +47,7 @@ class ServiceBusMessage(object):  # pylint: disable=too-many-public-methods,too-
     """A Service Bus Message.
 
     :param body: The data to send in a single message.
-    :type body: Union[str, bytes]
+    :type body: Optional[Union[str, bytes]]
 
     :keyword dict application_properties: The user defined properties on the message.
     :keyword str session_id: The session identifier of the message for a sessionful entity.
@@ -75,7 +77,7 @@ class ServiceBusMessage(object):  # pylint: disable=too-many-public-methods,too-
     """
 
     def __init__(self, body, **kwargs):
-        # type: (Union[str, bytes], Any) -> None
+        # type: (Optional[Union[str, bytes]], Any) -> None
         # Although we might normally thread through **kwargs this causes
         # problems as MessageProperties won't absorb spurious args.
         self._encoding = kwargs.pop("encoding", 'UTF-8')
@@ -109,11 +111,20 @@ class ServiceBusMessage(object):  # pylint: disable=too-many-public-methods,too-
         return str(self.message)
 
     def _build_message(self, body):
+        def _validate_body_type(body, allow_none=False):
+            if not (isinstance(body, (six.string_types, six.binary_type)) or (body is None and allow_none)):
+                raise TypeError("ServiceBusMessage body must be a string, bytes, None, or a list of string or bytes "
+                                "elements. Got instead: {}".format(type(body) if allow_none else "[... {} ...]"
+                                                                                                 .format(type(body))))
+
         if isinstance(body, list) and body:  # TODO: This only works for a list of bytes/strings
+            _validate_body_type(body[0])
             self.message = uamqp.Message(body[0], properties=self._amqp_properties, header=self._amqp_header)
             for more in body[1:]:
+                _validate_body_type(more)
                 self.message._body.append(more)  # pylint: disable=protected-access
         else:
+            _validate_body_type(body, allow_none=True)
             self.message = uamqp.Message(body, properties=self._amqp_properties, header=self._amqp_header)
 
     def _set_message_annotations(self, key, value):
@@ -267,7 +278,7 @@ class ServiceBusMessage(object):  # pylint: disable=too-many-public-methods,too-
 
     @property
     def body(self):
-        # type: () -> Union[bytes, Iterable[bytes]]
+        # type: () -> Optional[Union[bytes, Iterable[bytes]]]
         """The body of the Message.
 
         :rtype: bytes or Iterable[bytes]