+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 7.0.0b9 (Unreleased)
 
 **Breaking Changes**
+
 * `ServiceBusSender` and `ServiceBusReceiver` are no more reusable and will raise `ValueError` when trying to operate on a closed handler.
+* `ServiceBusMessage` will now raise a `TypeError` when provided an invalid body type.  Valid bodies are strings, bytes, None, or a list of string or bytes items.
 
 ## 7.0.0b8 (2020-11-05)